Oppdatert organisasjon hooks
