+++ apps/dolly-frontend/src/main/js/src/Store.js
@@ -10,7 +10,6 @@ import bestillingReducer from './ducks/bestilling'
 import kodeverkReducer from './ducks/kodeverk'
 import varslingerReducer from './ducks/varslinger'
 import finnPersonReducer from './ducks/finnPerson'
-import organisasjonReducer from './ducks/organisasjon'
 import redigertePersonerReducer from './ducks/redigertePersoner'
 import commonReducer from '~/ducks/common'
 import { createReduxHistoryContext, LOCATION_CHANGE } from 'redux-first-history'
@@ -53,7 +52,6 @@ const rootReducer = () =>
 		kodeverk: kodeverkReducer,
 		varslinger: varslingerReducer,
 		finnPerson: finnPersonReducer,
-		organisasjon: organisasjonReducer,
 		redigertePersoner: redigertePersonerReducer,
 		fasteData: fasteDataReducer,
 	})
