+++ apps/dolly-frontend/src/main/js/src/components/fagsystem/brregstub/form/partials/EgneOrganisasjoner.tsx
@@ -1,4 +1,4 @@
-import React, { useEffect, useState } from 'react'
+import React from 'react'
 import Loading from '~/components/ui/loading/Loading'
 import _get from 'lodash/get'
 import { FormikProps } from 'formik'
@@ -6,6 +6,7 @@ import { Adresse, Organisasjon } from '~/service/services/organisasjonforvalter/
 import { Alert } from '@navikt/ds-react'
 import { useCurrentBruker } from '~/utils/hooks/useBruker'
 import { EgneOrgSelect } from '~/components/ui/form/inputs/select/EgneOrgSelect'
+import { useOrganisasjoner } from '~/utils/hooks/useOrganisasjoner'
 
 interface OrgProps {
 	path: string
@@ -13,9 +14,6 @@ interface OrgProps {
 	formikBag: FormikProps<{}>
 	handleChange: (event: React.ChangeEvent<any>) => void
 	warningMessage?: string
-	isLoading: boolean
-	hentOrganisasjoner: Function
-	organisasjoner: Organisasjon[]
 	filterValidEnhetstyper?: boolean
 }
 
@@ -31,32 +29,75 @@ export const getAdresseWithAdressetype = (adresser: Adresse[], adressetype: stri
 		}))
 }
 
+const addAlleVirksomheter = (virksomheter: Organisasjon[], organisasjoner: Organisasjon[]) => {
+	for (let org of organisasjoner) {
+		virksomheter.push(org)
+		if (org.underenheter && org.underenheter.length > 0) {
+			addAlleVirksomheter(virksomheter, org.underenheter)
+		}
+	}
+}
+
+const getJuridiskEnhet = (orgnr: string, enheter: Organisasjon[]) => {
+	for (const enhet of enheter) {
+		if (enhet.underenheter && enhet.underenheter.length > 0) {
+			for (const underenhet of enhet.underenheter) {
+				if (underenhet.organisasjonsnummer === orgnr) {
+					return enhet.organisasjonsnummer
+				}
+				const juridisk: string = getJuridiskEnhet(orgnr, enhet.underenheter)
+				if (juridisk !== '') {
+					return juridisk
+				}
+			}
+		}
+	}
+	return ''
+}
+
+const getEgneOrganisasjoner = (organisasjoner: Organisasjon[]) => {
+	if (!organisasjoner) {
+		return []
+	}
+	const egneOrg: Organisasjon[] = []
+	addAlleVirksomheter(egneOrg, organisasjoner)
+
+	return egneOrg.map((org: Organisasjon) => {
+		const fAdresser = getAdresseWithAdressetype(org.adresser, 'FADR')
+		const pAdresser = getAdresseWithAdressetype(org.adresser, 'PADR')
+		const juridiskEnhet = getJuridiskEnhet(org.organisasjonsnummer, organisasjoner)
+		return {
+			value: org.organisasjonsnummer,
+			label: `${juridiskEnhet ? '   ' : ''}${org.organisasjonsnummer} (${org.enhetstype}) - ${
+				org.organisasjonsnavn
+			} ${juridiskEnhet ? '' : '   '}`,
+			orgnr: org.organisasjonsnummer,
+			navn: org.organisasjonsnavn,
+			enhetstype: org.enhetstype,
+			forretningsAdresse: fAdresser.length > 0 ? fAdresser[0] : null,
+			postAdresse: pAdresser.length > 0 ? pAdresser[0] : null,
+			juridiskEnhet: juridiskEnhet,
+		}
+	})
+}
+
 export const EgneOrganisasjoner = ({
 	path,
 	label,
 	formikBag,
 	handleChange,
 	warningMessage,
-	isLoading,
-	hentOrganisasjoner,
-	organisasjoner,
 	filterValidEnhetstyper,
 }: OrgProps) => {
-	const [error, setError] = useState(false)
 	const {
 		currentBruker: { brukerId },
 	} = useCurrentBruker()
 
-	const harEgneOrganisasjoner = organisasjoner && organisasjoner.length > 0
-	const validEnhetstyper = ['BEDR', 'AAFY']
+	const { organisasjoner, loading, error } = useOrganisasjoner(brukerId)
+	const egneOrganisasjoner = getEgneOrganisasjoner(organisasjoner)
 
-	useEffect(() => {
-		if (!organisasjoner) {
-			hentOrganisasjoner(brukerId).catch(() => {
-				setError(true)
-			})
-		}
-	}, [])
+	const harEgneOrganisasjoner = egneOrganisasjoner && egneOrganisasjoner.length > 0
+	const validEnhetstyper = ['BEDR', 'AAFY']
 
 	const getFilteredOptions = (organisasjoner: Organisasjon[]) => {
 		return organisasjoner
@@ -74,7 +115,7 @@ export const EgneOrganisasjoner = ({
 
 	return (
 		<>
-			{isLoading && <Loading label="Laster organisasjoner" />}
+			{loading && <Loading label="Laster organisasjoner" />}
 			{error && (
 				<Alert variant={'warning'}>
 					Noe gikk galt med henting av egne organisasjoner! Prøv på nytt, velg et annet alternativ
@@ -82,7 +123,7 @@ export const EgneOrganisasjoner = ({
 				</Alert>
 			)}
 			{!harEgneOrganisasjoner &&
-				!isLoading &&
+				!loading &&
 				!error &&
 				(warningMessage ? (
 					warningMessage
@@ -96,7 +137,9 @@ export const EgneOrganisasjoner = ({
 				<EgneOrgSelect
 					name={path}
 					label={label ? label : 'Organisasjonsnummer'}
-					options={filterValidEnhetstyper ? getFilteredOptions(organisasjoner) : organisasjoner}
+					options={
+						filterValidEnhetstyper ? getFilteredOptions(egneOrganisasjoner) : egneOrganisasjoner
+					}
 					size="xlarge"
 					onChange={handleChange}
 					value={_get(formikBag.values, path)}
